feat: use refine List component on Contact list page
