+++ examples/app-crm/src/routes/contacts/index.ts
@@ -1,4 +1,4 @@
-export * from "./wrapper";
+export * from "./list";
 export * from "./show";
 export * from "./create";
 export * from "./edit";
