--- examples/app-crm/src/routes/contacts/wrapper/index.module.css
@@ -1,13 +1,0 @@
-.container {
-    position: relative;
-    .header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 2.5rem;
-    }
-    .search {
-        display: flex;
-    }
-
-}
